segment fault na amtriz

+++ include/No.h
@@ -11,22 +11,22 @@ private:
     int idNo;
     int demanda;
     bool deposito = false;
-    int x;
-    int y;
+    float x;
+    float y;
     No *proxNo = nullptr;
     Aresta *primeiraAresta = nullptr;
     Aresta *ultimaAresta = nullptr;
 
 public:
 
-    No(int idNo, int x, int y);
+    No(int idNo, float x, float y);
     ~No();
 
     int getIdNo(){return idNo;};
     int getDemanda(){return demanda;};
     int getDeposito(){return deposito;};
-    int getX(){return x;};
-    int getY(){return y;};
+    float getX(){return x;};
+    float getY(){return y;};
     No * getProxNo(){return proxNo;};
     Aresta * getPrimeiraAresta(){return primeiraAresta;}
     Aresta * getUltimaAresta(){return ultimaAresta;}
@@ -34,7 +34,7 @@ public:
     void setProxNo(No *no);
     void setDemanda(int demanda);
     void setDeposito();
-    Aresta * insertAresta(int destino, int peso);
+    Aresta * insertAresta(int destino, float peso);
 };
 
 #endif // NO_H
